art2/requirements.txt
numpy==1.19.3
pandas
astar_art==0.0.1
python-mnist==0.7